Start of Website Update
The start of a large update to my website. This including the inclusion of my more recent projects. With the aim to give each project its own breakdown of systems whilst showcasing its overall gameplay.

Multiplayer Project -
Each element of the project has been described with a few placeholder photos/videos. These are to be replaced with more polished content.

Future Combat Simulator -
This has the showreel video included but now needs a breakdown of the systems putting on to the website.

About Me -
This is to be reworked and updated to show my new job experience.